JPA entity 수정
