+++ backend/src/main/kotlin/com/swpp/footprinter/domain/place/model/Place.kt
@@ -1,7 +1,7 @@
-package com.swpp.footprinter.core.domain.place.model
+package com.swpp.footprinter.domain.place.model
 
-import com.swpp.footprinter.core.common.model.BaseEntity
-import com.swpp.footprinter.core.domain.footprint.model.Footprint
+import com.swpp.footprinter.common.model.BaseEntity
+import com.swpp.footprinter.domain.footprint.model.Footprint
 import javax.persistence.Column
 import javax.persistence.Entity
 import javax.persistence.OneToMany
@@ -21,4 +21,4 @@ class Place(
     @OneToMany(mappedBy = "place")
     val footprints: List<Footprint>,
 
-) : BaseEntity()
+    ) : BaseEntity()
